requirements.txt

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
email-validator>=2.0.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Payments
stripe>=7.0.0

# Google Cloud
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0
google-cloud-firestore>=2.14.0
google-cloud-bigquery>=3.13.0
google-cloud-secret-manager>=2.16.0
google-generativeai>=0.3.0

# Monitoring
ddtrace>=2.4.0
datadog>=0.47.0

# Python 3.9 compatibility
importlib-metadata>=6.0.0

